pyproject.toml
[tool.poetry]
name = "genomics_final_project"
version = "0.1.0"
description = ""
authors = ["Jeremy Fisher", "Simon Levine-Gottreich", "Romel Baral"]

[tool.poetry.dependencies]
python = "3.7.6"
pandas = "^1.0.3"
numpy = "^1.18.2"
biopython = "^1.76"
tqdm = "^4.45.0"
tensorflow = "2.0.1"

[tool.poetry.dev-dependencies]
nb-clean = "^1.4.1"
pytest = "^5.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
